Src/WebServerApp/requirements.txt
flask==2.3.2
numpy==1.24.3
pybase64==1.2.3
redis==4.6.0
PyYAML==5.4.1
httplib2==0.20.2